Написаны тесты для UserService
